requirements.txt
fastapi
python-telegram-bot==20.7
uvicorn
pydantic
